+++ docs/gen_paramters_doc.py
@@ -10,40 +10,40 @@ Parameters
 
 About Paramters
 ---------------
-The CCHDO works frequently with many paramters common in hydrography.
-Provided here is a description of many common paramters encountered in exchange files.
+The CCHDO works frequently with many parameters common in hydrography.
+Provided here is a description of many common parameters encountered in exchange files.
 
 .. warning::
-  This list may not contain every paramter which may be encountered in an exchange file.
-  CCHDO is working on providing a machine readable list of ALL paramters which may be encountered in all files.
-  This list will include paramters which CCHDO lacks a descirption for (known unknown parmaters).
-  When this list becomes available, we intend to tie the generation of this paramter list with the machine readable list of paramters.
+  This list may not contain every parameter which may be encountered in an exchange file.
+  CCHDO is working on providing a machine readable list of ALL parameters which may be encountered in all files.
+  This list will include parameters which CCHDO lacks a descirption for (known unknown parmaters).
+  When this list becomes available, we intend to tie the generation of this parameter list with the machine readable list of parameters.
 
-  Until that time, paramters may appear in exchange formats from the CCHDO which are not documented here.
-  No undocumented paramter or field will cause data integrity or usefulness issues (i.e. all undocumented paramters may be safely ignored).
+  Until that time, parameters may appear in exchange formats from the CCHDO which are not documented here.
+  No undocumented parameter or field will cause data integrity or usefulness issues (i.e. all undocumented parameters may be safely ignored).
 
 Definitions
 ^^^^^^^^^^^
-Provided with each paramter is a set of information in a table, the information included in that table should be interpreted as follows:
+Provided with each parameter is a set of information in a table, the information included in that table should be interpreted as follows:
 
 * Units: 
-    These are the common units encountered for this paramter as it will appear in the exchange document itself.
+    These are the common units encountered for this parameter as it will appear in the exchange document itself.
     The special units of "None" means the field will be either blank or contain only whitespace.
 * Data Type: 
-    Specifies the allowed type of data in the data records for this paramter.
+    Specifies the allowed type of data in the data records for this parameter.
     There are three types of data, string (str), integers (int), and decimal.
     String data types may be any printing charicter except a comma ``,`` which is the field seperator.
     Integer data types may only contain numbers without a decimal point, quality flags are usually integers.
     Decimal data types may be any real number (including an integer) and may include decimal point, the precision is not specified.
 * Quality Flags: 
-    Specifies which set of quality flag definitions should be used to interpret a quality flag column for this paramter (if present).
+    Specifies which set of quality flag definitions should be used to interpret a quality flag column for this parameter (if present).
     Current quality flags are: :ref:`bottle <Bottle Quality Codes>`, :ref:`water <Water Quality Codes>`, :ref:`ctd <CTD Quality Codes>`.
     See the :ref:`Quality Codes` section for more information
 
 Common Parameters
 -----------------
 This section was generated automatically from a 
-:download:`machine readable list of paramters <metadata/parameters.json>`.
+:download:`machine readable list of parameters <metadata/parameters.json>`.
 
 .. hlist::
   :columns: 3