+++ docs/changelog.rst
@@ -11,7 +11,7 @@ Changelog
 
 2014-08-18
 ----------
-* Less restrictive paramter names and units.
+* Less restrictive parameter names and units.
 * Specify how numerical data should appear.
 * CCHDO now keeps numerical precsion of data found in files